requirements.txt
python-telegram-bot==13.7
urllib3==1.26.6
six==1.16.0
filetype==1.2.0
GitPython==3.1.43